Rename sig2str.rs to str2sig.rs.
The old name makes little sense now that the `sig2str` function is
gone.
