+++ rust_src/src/lib.rs
@@ -28,7 +28,7 @@ mod symbols;
 mod character;
 mod base64;
 mod crypto;
-mod sig2str;
+mod str2sig;
 
 use remacs_sys::Lisp_Subr;
 
@@ -64,7 +64,7 @@ pub use crypto::sha1_ctx::sha1_process_bytes;
 pub use crypto::sha1_ctx::sha1_finish_ctx;
 
 // Used in process.c
-pub use sig2str::str2sig;
+pub use str2sig::str2sig;
 
 // These need to be exported as marker.c depends upon them.
 pub use marker::CHECK_MARKER;
