+++ rust_src/src/str2sig.rs
